# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "embeddings-rs",
    "description": "Embedding sidecar for nomic-embed-text-v1.5 (768-dim). Drop-in for the Python `embeddings` service.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.1"
  },
  "paths": {
    "/embed": {
      "post": {
        "tags": [
          "embeddings"
        ],
        "summary": "Embed a single text into a 768-dim vector.",
        "operationId": "handle_embed",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Embedding vector",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Inference failed"
          },
          "503": {
            "description": "No usable inference lanes remain"
          }
        }
      }
    },
    "/embed-batch": {
      "post": {
        "tags": [
          "embeddings"
        ],
        "summary": "Embed multiple texts in one call.",
        "operationId": "handle_embed_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Embedding vectors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedBatchResponse"
                }
              }
            }
          },
          "500": {
            "description": "Inference failed"
          },
          "503": {
            "description": "No usable inference lanes remain"
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "embeddings"
        ],
        "summary": "Liveness + model info + real inference-lane capacity.",
        "description": "Reports the pool's ACTUAL state and fails with 503 once no lane can serve a\nrequest. The previous version returned a hardcoded `\"healthy\"` and touched\nneither the pool nor inference — which is why `docker ps` reported healthy\nfor the seven days this service was completely dead.",
        "operationId": "handle_health",
        "responses": {
          "200": {
            "description": "Healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "503": {
            "description": "No usable inference lanes remain"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EmbedBatchRequest": {
        "type": "object",
        "required": [
          "texts"
        ],
        "properties": {
          "task_type": {
            "type": "string",
            "default": "search_document",
            "example": "search_document"
          },
          "texts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Texts to embed. Processed in token-bounded sub-batches internally, so a\nlarge request (or one with a very long text) is handled without a memory\nspike; vectors are returned in input order."
          }
        }
      },
      "EmbedBatchResponse": {
        "type": "object",
        "required": [
          "vectors",
          "count",
          "dimensions"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "dimensions": {
            "type": "integer",
            "minimum": 0
          },
          "vectors": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "number",
                "format": "float"
              }
            }
          }
        }
      },
      "EmbedRequest": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "task_type": {
            "type": "string",
            "description": "nomic task prefix; prepended as `\"{task_type}: {text}\"` before encoding.",
            "default": "search_document",
            "example": "search_document"
          },
          "text": {
            "type": "string",
            "description": "Text to embed.",
            "example": "The dragon coiled around the obsidian spire."
          }
        }
      },
      "EmbedResponse": {
        "type": "object",
        "required": [
          "vector",
          "dimensions"
        ],
        "properties": {
          "dimensions": {
            "type": "integer",
            "description": "Always 768 for nomic-embed-text-v1.5.",
            "minimum": 0
          },
          "vector": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            },
            "description": "Raw (un-normalized) masked-mean-pooled embedding; `|v|` ~= 22."
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": [
          "status",
          "model",
          "dimensions",
          "lanes_total",
          "lanes_available"
        ],
        "properties": {
          "dimensions": {
            "type": "integer",
            "minimum": 0
          },
          "lanes_available": {
            "type": "integer",
            "description": "Lanes that can still serve a request. Equals `lanes_total` in a healthy\nprocess; a lower number means capacity was permanently lost and is worth\nalerting on. When this reaches 0 the endpoint returns **503**, because a\nliveness probe that passes while every request fails is worse than none.",
            "minimum": 0
          },
          "lanes_total": {
            "type": "integer",
            "description": "Inference lanes the pool was built with (`EMBEDDING_POOL_SIZE`).",
            "minimum": 0
          },
          "model": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "embeddings",
      "description": "Text embedding endpoints"
    }
  ]
}